requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
deta==1.1.0
idna==3.4
pydotenv==0.0.7
python-dotenv==0.21.0
requests==2.28.1
soupsieve==2.3.2.post1
urllib3==1.26.12
